complete-code/requirements.txt
mlflow
scikit-learn
pandas
numpy
pyyaml
fastapi
uvicorn
dvc[s3]
scipy
